some errors
I think in Transaction method --> getCategoryName is error, at least i had long time with it. 
I make screenshotes of it, and included in that class comment out under method code I use to fix it while was doing table view... Pavel check it. 
Also check ohter 4 pictures, and methdo in MenuuActionProgrammableHandler.
Every thing is working as it is in doParsOFXFileProcessing1() method (with Transaction mehtod as now comment out) but doParsOFXFileProcessing2() not parsing it, check pictures.... 
I was try to debugg it for some times but unsuccessfully. But still thinking how to make ttable clickable  and sortable. Probably willl make it work soon.